+++ src/arg_parser/pscan_parser.rs
@@ -1,4 +1,6 @@
-use crate::prelude::{Parser, Ipv4Addr};
+pub use std::net::IpAddr;
+pub use clap::Parser;
+
 
 
 #[derive(Parser)]
