+++ src/engines/portscan.rs
@@ -1,9 +1,9 @@
-use crate::prelude::{
-    BTreeSet, Duration, thread, SliceRandom, PortScanArgs, Parser,
-    PacketBuilder, PacketDissector, PacketSender, PacketSniffer,
-    get_host_name, display_progress, display_error_and_exit
-};
-
+use std::{thread, time::Duration};
+use clap::Parser;
+use rand::seq::SliceRandom;
+use arg_parser::pscan_parser::PortScanArgs;
+use crate::packets::{PacketBuilder, PacketDissector, PacketSender, PacketSniffer};
+use crate::utils::{PortGenerator, display_error_and_exit, display_progress, get_host_name};
 
 
 
@@ -66,70 +66,6 @@ impl PortScanner {
             display_progress(format!("Packet sent to port: {:<5} - {:<15}", port, ip));
             thread::sleep(Duration::from_secs_f32(0.02));
         }
-    }
-
-
-
-    fn get_ports(&self) -> Vec<u16> {
-        let mut ports_vec: Vec<u16> = match self.args.ports {
-            Some(_) => self.generate_specified_ports(),
-            None   => (1..=100).collect(),
-        };
-
-        if self.args.random {
-            Self::shuffle_ports(&mut ports_vec);
-        }
-        
-        ports_vec
-    }
-
-
-
-    fn generate_specified_ports(&self) -> Vec<u16> {
-        let mut ports: BTreeSet<u16> = BTreeSet::new();
-        let parts: Vec<&str>         = self.args.ports.as_ref().clone().unwrap().split(",").collect();
-        
-        for part in parts {
-            if part.contains("-") {
-                ports.extend(Self::get_port_range(part.to_string()));
-            } else {
-                ports.insert(Self::validate_port(part.to_string()));
-            }
-        }
-
-        let ports_vec: Vec<u16> = ports.into_iter().collect();
-        ports_vec
-    }
-
-
-
-    fn get_port_range(port_range: String) -> Vec<u16> {
-        let parts: Vec<&str> = port_range.split("-").collect();
-        let start: u16       = Self::validate_port(parts[0].to_string());
-        let end: u16         = Self::validate_port(parts[1].to_string());
-        
-        if start >= end {
-            display_error_and_exit(format!("Invalid range format {}-{}", start, end));
-        }
-
-        (start..=end).collect()
-    }
-
-
-
-    fn validate_port(port_str: String) -> u16 {
-        let port: u16 = port_str.parse().unwrap_or_else(|_| {
-            display_error_and_exit(format!("Invalid port: {}", port_str));
-        });
-
-        port
-    }
-
-
-
-    fn shuffle_ports(ports_vec: &mut Vec<u16>) {
-        let mut rng = rand::thread_rng();
-        ports_vec.shuffle(&mut rng);
     }
