+++ src/packets/pkt_sniffer.rs
@@ -1,6 +1,12 @@
-use crate::prelude::{
-    thread, Arc, Mutex, AtomicBool, Ordering, Device, Capture, get_default_iface_ip, get_network
+use std::{
+    thread,
+    sync::{
+        Arc, Mutexsync,
+        atomic::AtomicBool, Ordering,
+    },
 };
+use pcap::{Device, Capture};
+use crate::utils::{get_default_iface_ip, get_network};
