--- src/prelude.rs
@@ -1,55 +1,0 @@
-pub use std::{
-    collections::{HashMap, BTreeSet},
-    env,
-    io::{self, Write},
-    net::{Ipv4Addr, IpAddr},
-    sync::{Arc, Mutex},
-    sync::atomic::{AtomicBool, Ordering},
-    thread,
-    time::Duration,
-};
-
-
-
-
-pub use clap::Parser;
-pub use ipnet::{Ipv4AddrRange, Ipv4Net};
-pub use dns_lookup::lookup_addr;
-pub use etherparse::{SlicedPacket, InternetSlice, LinkSlice};
-pub use netdev::interface::get_default_interface;
-pub use pcap::{Device, Capture};
-pub use rand::{Rng, seq::SliceRandom};
-
-pub use pnet::{
-    packet::{
-        ip::{IpNextHeaderProtocols, IpNextHeaderProtocol},
-        ipv4::{MutableIpv4Packet, checksum as ip_checksum},
-        tcp::{MutableTcpPacket, TcpFlags, ipv4_checksum as tcp_checksum},
-    },
-    transport::{transport_channel, TransportChannelType::Layer3, TransportSender},
-};
-
-
-
-
-pub use crate::arg_parser::{
-    pscan_parser::PortScanArgs,
-};
-
-pub use crate::engines::{
-    netmap::NetworkMapper,
-    portscan::PortScanner
-};
-
-pub use crate::packets::{
-    pkt_builder::PacketBuilder,
-    pkt_dissector::PacketDissector,
-    pkt_sender::PacketSender,
-    pkt_sniffer::PacketSniffer,
-};
-
-pub use crate::utils::{
-    displays::{display_error_and_exit, display_progress},
-    iface_info::{get_default_iface_info, get_default_iface_ip, get_network},
-    network_info::get_host_name
-};
