+++ src/utils/mod.rs
@@ -1,0 +1,9 @@
+pub mod displays;
+pub mod iface_info;
+pub mod network_info;
+pub mod port_generator;
+
+pub use displays::{display_error_and_exit, display_progress};
+pub use iface_info::{get_default_iface_info, get_default_iface_ip, get_network};
+pub use network_info::get_host_name;
+pub use port_generator::PortGenerator;
