+++ src/utils/network_info.rs
@@ -1,4 +1,5 @@
-use crate::prelude::{IpAddr, lookup_addr};
+use std::net::IpAddr,
+use dns_lookup::lookup_addr,
